deployed CR_ConditioningMixer
