+++ dev_node_mappings.py
@@ -9,12 +9,11 @@ from .dev_nodes.dev_schedulers import *
 from .dev_nodes.dev_xygrid import *
 
 DEV_NODE_CLASS_MAPPINGS = {
-    ### Dev Nodes
+    ### XY Dev Nodes
     "CR XYZ List": CR_XYZList,    
     "CR XYZ Interpolate": CR_XYZInterpolate,    
     "CR XYZ Index": CR_XYZIndex,
-    "CR Conditioning Mixer": CR_ConditioningMixer,  
-    ### Graphics_Dev Nodes
+    ### Graphics Dev Nodes
     "CR Overlay Transparent Image": CR_OverlayTransparentImage,
     "CR Multi-Panel Meme Template": CR_MultiPanelMemeTemplate,
     "CR Popular Meme Templates": CR_PopularMemeTemplates,    
@@ -24,7 +23,7 @@ DEV_NODE_CLASS_MAPPINGS = {
     "CR Add Annotation": CR_AddAnnotation,    
     "CR 3D Polygon": CR_3DPolygon,
     "CR 3D Solids": CR_3DSolids,
-    ### Workflow
+    ### Workflow Dev Nodes
     "CR Job List": CR_JobList,
     "CR Job Scheduler": CR_JobScheduler,
     "CR Check Job Complete": CR_CheckJobComplete,
@@ -73,7 +72,6 @@ DEV_NODE_DISPLAY_NAME_MAPPINGS = {
     "CR Spawn Workflow Instance": "CR Spawn Workflow Instance (Prototype)",    
     "CR Job Current Frame": "CR Job Current Frame (Prototype)",
     "CR Load Workflow": "CR Load Workflow (Prototype)",
-    "CR Conditioning Mixer": "CR Conditioning Mixer (Dev)",
     ### Animation Dev Nodes
     "CR Prompt Weight Scheduler": "CR Prompt Weight Scheduler (Dev)",    
     "CR Load Scheduled ControlNets": "CR Load Scheduled ControlNets (Dev)",
