+++ dev_nodes/dev_nodes.py
@@ -3,91 +3,16 @@
 # for ComfyUI                                           https://github.com/comfyanonymous/ComfyUI                                               
 #---------------------------------------------------------------------------------------------------------------------#
 
-import numpy as np
-import torch
-import os 
 from ..categories import icons
 
 #---------------------------------------------------------------------------------------------------------------------#
-class CR_ConditioningMixer:
 
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        mix_methods = ["Combine", "Average", "Concatenate"]
-        
-        return {"required":
-                    {"conditioning_1": ("CONDITIONING", ),
-                     "conditioning_2": ("CONDITIONING", ),
-                     "average_method_strength": ("FLOAT", {"default": 1.0, "min": 0.0, "max": 1.0, "step": 0.01}),
-                     "mix_method": (mix_methods, ),
-                     }
-                }
-
-    RETURN_TYPES = ("CONDITIONING", )
-    FUNCTION = "conditioning"
-    CATEGORY = icons.get("Comfyroll/Other")
-    
-    def conditioning(self, mix_method, conditioning_1, conditioning_2, average_method_strength):
-
-        conditioning_from = conditioning_1
-        conditioning_to = conditioning_2
-        conditioning_to_strength = average_method_strength
-    
-        if mix_method == "Combine":
-            return (conditioning_1 + conditioning_2, )
-
-        if mix_method == "Average":
-        
-            out = []
-
-            if len(conditioning_from) > 1:
-                print("Warning: ConditioningAverage conditioning_from contains more than 1 cond, only the first one will actually be applied to conditioning_to.")
-
-            cond_from = conditioning_from[0][0]
-            pooled_output_from = conditioning_from[0][1].get("pooled_output", None)
-
-            for i in range(len(conditioning_to)):
-                t1 = conditioning_to[i][0]
-                pooled_output_to = conditioning_to[i][1].get("pooled_output", pooled_output_from)
-                t0 = cond_from[:,:t1.shape[1]]
-                if t0.shape[1] < t1.shape[1]:
-                    t0 = torch.cat([t0] + [torch.zeros((1, (t1.shape[1] - t0.shape[1]), t1.shape[2]))], dim=1)
-
-                tw = torch.mul(t1, conditioning_to_strength) + torch.mul(t0, (1.0 - conditioning_to_strength))
-                t_to = conditioning_to[i][1].copy()
-                if pooled_output_from is not None and pooled_output_to is not None:
-                    t_to["pooled_output"] = torch.mul(pooled_output_to, conditioning_to_strength) + torch.mul(pooled_output_from, (1.0 - conditioning_to_strength))
-                elif pooled_output_from is not None:
-                    t_to["pooled_output"] = pooled_output_from
-
-                n = [tw, t_to]
-                out.append(n)
-            return (out,)
-
-        if mix_method == "Concatenate":
-        
-            out = []
-
-            if len(conditioning_from) > 1:
-                print("Warning: ConditioningConcat conditioning_from contains more than 1 cond, only the first one will actually be applied to conditioning_to.")
-
-            cond_from = conditioning_from[0][0]
-
-            for i in range(len(conditioning_to)):
-                t1 = conditioning_to[i][0]
-                tw = torch.cat((t1, cond_from),1)
-                n = [tw, conditioning_to[i][1].copy()]
-                out.append(n)
-            return (out, )
-            
 #---------------------------------------------------------------------------------------------------------------------#
 # MAPPINGS
 #---------------------------------------------------------------------------------------------------------------------#
 # For reference only, actual mappings are in __init__.py
 '''
 NODE_CLASS_MAPPINGS = {
-    "CR Conditioning Mixer":CR_ConditioningMixer,
 }
 '''
