+++ nodes/nodes.py
@@ -1,7 +1,7 @@
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
-# Comfyroll Custom Nodes by RockOfFire and Akatsuzi         https://github.com/RockOfFire/ComfyUI_Comfyroll_CustomNodes                             #
-# for ComfyUI                                               https://github.com/comfyanonymous/ComfyUI                                               #
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
+#---------------------------------------------------------------------------------------------------------------------#
+# Comfyroll Custom Nodes by RockOfFire and Akatsuzi       https://github.com/RockOfFire/ComfyUI_Comfyroll_CustomNodes
+# for ComfyUI                                             https://github.com/comfyanonymous/ComfyUI
+#---------------------------------------------------------------------------------------------------------------------#
 
 import torch
 import numpy as np
@@ -17,8 +17,7 @@ from ..categories import icons
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.realpath(__file__)), "comfy"))
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#           
-# This node has a bunch of default picture widths and heights as well as being able to choose them yourself.
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_AspectRatioSD15:
     def __init__(self):
         pass
@@ -76,9 +75,7 @@ class CR_AspectRatioSD15:
         else:
             return(width, height, upscale_factor, batch_size,)  
             
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
-
-#Legacy Node. This node was an attempt at making a save and preview image node into one.
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_ImageOutput:
     def __init__(self):
         self.output_dir = folder_paths.get_output_directory()
@@ -164,8 +161,7 @@ class CR_ImageOutput:
         return { "ui": { "images": results }, "result": (trigger,) }
 
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
-# This node will give you the multiple of the value within the interger parameter.
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_IntegerMultipleOf:
     def __init__(self):
         pass
@@ -189,8 +185,7 @@ class CR_IntegerMultipleOf:
         integer = integer * multiple
         return (int(integer), )
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
-# This node is for making seeds
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_Seed:
     def __init__(self):
         pass
@@ -212,8 +207,7 @@ class CR_Seed:
     def seedint(seed):
         return seed,
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
-# Sets the latent batch size
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_LatentBatchSize:
 
     def __init__(self):
@@ -244,8 +238,7 @@ class CR_LatentBatchSize:
             'samples': torch.cat(sample_list),
         }, )
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
-#Text for prompts
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_PromptText:
     @classmethod
     def INPUT_TYPES(s):
@@ -262,7 +255,7 @@ class CR_PromptText:
     def get_value(self, prompt):
         return (prompt,)
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_SplitString:
 
     @classmethod
@@ -287,7 +280,7 @@ class CR_SplitString:
 
         return (string_1, string_2, string_3, string_4)
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
+#---------------------------------------------------------------------------------------------------------------------#
 class CR_Value:
 
     @classmethod
@@ -304,9 +297,82 @@ class CR_Value:
     def get_value(self, value):
         return (float(value), int(value), )
 
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
+#---------------------------------------------------------------------------------------------------------------------#
+class CR_ConditioningMixer:
+
+    @classmethod
+    def INPUT_TYPES(s):
+    
+        mix_methods = ["Combine", "Average", "Concatenate"]
+        
+        return {"required":
+                    {"conditioning_1": ("CONDITIONING", ),
+                     "conditioning_2": ("CONDITIONING", ),      
+                     "mix_method": (mix_methods, ),
+                     "average_strength": ("FLOAT", {"default": 0.5, "min": 0.0, "max": 1.0, "step": 0.01}),
+                     }
+                }
+
+    RETURN_TYPES = ("CONDITIONING", )
+    FUNCTION = "conditioning"
+    CATEGORY = icons.get("Comfyroll/Other")
+    
+    def conditioning(self, mix_method, conditioning_1, conditioning_2, average_strength):
+
+        conditioning_from = conditioning_1
+        conditioning_to = conditioning_2
+        conditioning_to_strength = average_strength
+    
+        if mix_method == "Combine":
+            return (conditioning_1 + conditioning_2, )
+
+        if mix_method == "Average":
+        
+            out = []
+
+            if len(conditioning_from) > 1:
+                print("Warning: ConditioningAverage conditioning_from contains more than 1 cond, only the first one will actually be applied to conditioning_to.")
+
+            cond_from = conditioning_from[0][0]
+            pooled_output_from = conditioning_from[0][1].get("pooled_output", None)
+
+            for i in range(len(conditioning_to)):
+                t1 = conditioning_to[i][0]
+                pooled_output_to = conditioning_to[i][1].get("pooled_output", pooled_output_from)
+                t0 = cond_from[:,:t1.shape[1]]
+                if t0.shape[1] < t1.shape[1]:
+                    t0 = torch.cat([t0] + [torch.zeros((1, (t1.shape[1] - t0.shape[1]), t1.shape[2]))], dim=1)
+
+                tw = torch.mul(t1, conditioning_to_strength) + torch.mul(t0, (1.0 - conditioning_to_strength))
+                t_to = conditioning_to[i][1].copy()
+                if pooled_output_from is not None and pooled_output_to is not None:
+                    t_to["pooled_output"] = torch.mul(pooled_output_to, conditioning_to_strength) + torch.mul(pooled_output_from, (1.0 - conditioning_to_strength))
+                elif pooled_output_from is not None:
+                    t_to["pooled_output"] = pooled_output_from
+
+                n = [tw, t_to]
+                out.append(n)
+            return (out,)
+
+        if mix_method == "Concatenate":
+        
+            out = []
+
+            if len(conditioning_from) > 1:
+                print("Warning: ConditioningConcat conditioning_from contains more than 1 cond, only the first one will actually be applied to conditioning_to.")
+
+            cond_from = conditioning_from[0][0]
+
+            for i in range(len(conditioning_to)):
+                t1 = conditioning_to[i][0]
+                tw = torch.cat((t1, cond_from),1)
+                n = [tw, conditioning_to[i][1].copy()]
+                out.append(n)
+            return (out, )
+            
+#---------------------------------------------------------------------------------------------------------------------#
 # MAPPINGS
-#---------------------------------------------------------------------------------------------------------------------------------------------------#
+#---------------------------------------------------------------------------------------------------------------------#
 # For reference only, actual mappings are in __init__.py
 '''
 NODE_CLASS_MAPPINGS = {
@@ -318,6 +384,7 @@ NODE_CLASS_MAPPINGS = {
     "CR Prompt Text":CR_PromptText,
     "CR Split String":CR_SplitString,
     "CR Value": CR_Value,
+    "CR Conditioning Mixer":CR_ConditioningMixer,    
 }
 '''
